Update: Added Cart view. Fix: Bug where all products inside the Product model were populating the cart if Add to cart was hit on just one. Added Promise class to fix issue. Temporary fix for now.

+++ models/cart.js
@@ -16,23 +16,79 @@ module.exports = class Cart {
 				cart = JSON.parse( fileContent )
 			}
 			// Analyze the cart => find existing products
-			const existingProductIndex = cart.products.findIndex(prod => prod.id === id);
-			const existingProduct = cart.products[existingProductIndex];
+			const existingProductIndex = cart.products.findIndex( prod => prod.id === id );
+			const existingProduct      = cart.products[ existingProductIndex ];
 			
-			if(existingProduct) {
+			// Add new products / increase quantity
+			
+			if( existingProduct ){
 				existingProduct.qty += 1;
 			} else {
 				const newProduct = { id: id, qty: 1 };
-				cart.products.push(newProduct);
+				cart.products.push( newProduct );
 			}
-			cart.totalPrice += +productPrice;
-			fs.writeFile(p, JSON.stringify(cart), err => {
-				console.log(err);
-			});
+			cart.totalPrice = ( cart.totalPrice || 0 ) + +productPrice;
+			fs.writeFile( p, JSON.stringify( cart ), err => {
+				console.log( err );
+			} );
 			
 		} )
 		
-		// Add new products / increase quantity
+	}
+	
+	static deleteProduct( id, productPrice ){
+		
+		// Fetch current cart
+		
+		fs.readFile( p, ( err, fileContent ) => {
+			
+			if( err ){
+				console.log( 'Error reading cart file', err )
+				return
+			}
+			
+			let cart = JSON.parse( fileContent )
+			// Find the product to delete
+			
+			const existingProductIndex = cart.products.findIndex( prod => prod.id === id )
+			
+			if( existingProductIndex !== -1 ){
+				
+				const existingProduct = cart.products[ existingProductIndex ]
+				
+				// Reduce the quantity
+				
+				existingProduct.qty -= 1;
+				cart.totalPrice = ( cart.totalPrice || 0 ) - +productPrice
+				
+				if( existingProduct.qty === 0 ){
+					// Remove the product
+					cart.products.splice( existingProductIndex, 1 )
+				}
+				
+				// Write updated cart back to the file
+				
+				fs.writeFile( p, JSON.stringify( cart ), err => {
+					console.log( err )
+				} )
+				
+			} else {
+				console.log( "Product not found." );
+			}
+			
+		} )
+	}
+	
+	static getCart( cb ){
+		fs.readFile( p, ( err, fileContent ) => {
+			const cart = JSON.parse( fileContent );
+			
+			if( err ){
+				cb( null );
+			} else {
+				cb( cart );
+			}
+		} )
 	}
 	
 }
